Minor fixes, added JavaDocs
