+++ hedera-node/hedera-app-spi/build.gradle.kts
@@ -20,7 +20,9 @@ plugins {
 description = "Hedera Application - SPI"
 
 dependencies {
-    api(libs.bundles.swirlds)
+    api(libs.swirlds.common)
+    api(libs.swirlds.jasperdb)
+    api(libs.swirlds.virtualmap)
 
     implementation(libs.hapi)
 }
