+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/QueryHandler.java
@@ -15,4 +15,5 @@
  */
 package com.hedera.node.app.spi;
 
+/** Marker interface for all handlers that query entities. */
 public interface QueryHandler {}
