+++ hedera-node/hedera-app/src/main/java/com/hedera/services/api/implementation/ServicesAccessor.java
@@ -19,6 +19,13 @@ import com.hedera.node.app.service.file.FileService;
 import com.hedera.node.app.service.token.CryptoService;
 import com.hedera.node.app.service.token.TokenService;
 
+/**
+ * A {@code ServiceAccessor} is used to pass all services to components via a single parameter.
+ *
+ * @param cryptoService a {@link CryptoService}
+ * @param fileService a {@link FileService}
+ * @param tokenService a {@link TokenService}
+ */
 public record ServicesAccessor(
         ServicesContext<CryptoService> cryptoService,
         ServicesContext<FileService> fileService,
